refactor(landing): consolidate window management and extract hooks - decompose LandingPage.tsx into specialized hooks (useItemActions, useChatManager, etc) - extract generic window helpers to src/lib/window-helpers.ts for global reuse - create useDiabloFocus hook to centralize focus detection - refactor ItemOverlayWidget to use event-driven window management - centralize window references to prevent synchronization issues
